checks if current track from web & device are same
Most recent error added to log should appear at top of the error log. Broadcast receiver now uses com.spotify.music.metadatachanged instead of com.spotify.music.active so that it  can see the uri of the track that just started playing. This is compared to the uri from the web api request to insure that both responses are in sync.